requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
aiohttp==3.9.1
openai==1.3.7
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
# RAG System dependencies
chromadb==0.4.22
sentence-transformers==2.2.2
torch==2.1.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0 